chore(shared): move hooks to their own module
