+++ packages/shared/src/helpers.ts
@@ -1,4 +1,3 @@
-import { useEffect, useRef, useState } from 'react'
 import { Lookup, Arrify } from './types.util'
 import * as G from './globals'
 
@@ -73,31 +72,3 @@ export const each = <T extends object, This>(
 
 export const toArray = <T>(a: T): Arrify<Exclude<T, void>> =>
   is.und(a) ? [] : is.arr(a) ? (a as any) : [a]
-
-// Explicit type annotation fixes TS2742 error.
-type UseOnce = (effect: React.EffectCallback) => void
-
-export const useOnce: UseOnce = effect => useEffect(effect, [])
-
-/** Return a function that re-renders this component, if still mounted */
-export const useForceUpdate = () => {
-  const update = useState<any>(0)[1]
-  const unmounted = useRef(false)
-  useOnce(() => () => {
-    unmounted.current = true
-  })
-  return () => {
-    if (!unmounted.current) {
-      update({})
-    }
-  }
-}
-
-/** Use a value from the previous render */
-export function usePrev<T>(value: T): T | undefined {
-  const prevRef = useRef<any>(undefined)
-  useEffect(() => {
-    prevRef.current = value
-  })
-  return prevRef.current
-}
