+++ packages/shared/src/index.ts
@@ -4,6 +4,7 @@ export { Globals }
 export * from './types'
 export * from './types.util'
 
+export * from './hooks'
 export * from './helpers'
 export * from './FrameLoop'
 export * from './createInterpolator'
